Setting both purge and config_replace to false.
Config settings that have this type of chnage should not be set to true by default. Let
the user decide if they want to be making these changes to their system - I removed my own
sudo access twice before I realized that it was this module doing so (Luckily was on a trash
AWS instance so I didn't mind).