use upstream (sudo.ws) sudo, remove ldap dependency
this commit changes the default sudo installation to
upstream (www.sudo.ws).

This also removes the dependency on ldap, as there are upstream packages
without ldap support.

the perzl.org package is ok, but i think it's better to use the upstream
version. we are also going to change the sudo installation to sudo.ws on
our AIX systems.