if config_replace is false use augeas to add #include_dir
on redhat 5 with sudo 1.7 there's no #include_dir option in
/etc/suoders.

if config_replace is not an option, just add the #include_dir option
with augeas and we are done. This is especially useful if there are
existing sudo entries in /etc/sudoer you do not want to overwrite.

this also includes the sudoers augeas lens from redhat 6 as there are
lot of cases not supported by the default redhat 5 lens. (e.g uppercase
usernames in /etc/suders and so on...).