Change the way visudo checks are done with 2 options.
The 2 options are $delete_on_error and $validate_single.

If $delete_on_error is true the configuration is deleted if visudo -c return
an error. If false the error will be only notified but the file will not be
deleted. Default is true

if $validate_single is true the file resource will run with validate_cmd to
test the single files and if there's an error the old file will not be
substituted or deleted. Default is false