Missing path depending on sudo package
If they are using a different sudo module on an os (say Solaris) and/or a different place to store binaries.  It might be even better to set the path as a variable and have the default be the bin/sbin.  Since path is declared in this exec, there's no other way to override.