Add Puppet binaries to secure_path
The `/opt/puppetlabs/bin` directory has been appended to the secure_path
for the sudoers template in this project since January, but the files
have not been updated. There is no real use for the template update then
since most sudoers files will be based on the static files, for real
systems at least. Therefore, this commit adds that Puppet binaries
directory to the secure_path for all the static sudoers files which
have a secure_path in them so that real systems get them and can simply
run `sudo puppet` instead of having to rely on using the full path.

Fixes saz/puppet-sudo#177.