New parameter suffix to main class
If optional parameter `suffix` is specified then the specified
suffix will be added to each file entry in the `/etc/sudoers.d`
directory.

The motivation here is to allow a configuration.

```puppet
class{'sudo':
  suffix => '_puppet',
  ignore => '*[!_puppet]',
}
sudo::conf{'admins':
  content => '...',
}
```
this will allow puppet to only purge the files that it created.
In particular sudo file entries which are dropped in via an RPM
will not be clobbered by the module. However puppet created
files will still be purged.