Add prefix parameter to prefix all sudoers.d entries
A previous attempt in

https://github.com/saz/puppet-sudo/pull/248

to allow only files created with the suffix `_puppet` to be
purged is faulty. In particular with

```
class{'sudo':
  suffix => '_puppet',
  purge_ignore => '*[!_puppet]',
}
```

Does not work as files names shorter than the length of string `_puppet` are
not ignored and so are deleted.

Switching to prefix

```
class{'sudo':
  prefix => 'puppet_',
  purge_ignore => '[!puppet_]*',
}
```

Gives the desired result.

```txt
irb(main):001:0> Dir.glob('[!puppet_]*')
=> ["short", "longfilename"]

So the files `short` and `longfilename` will be ignored correctly.